Moved sampler to another file
